refactor middleware and route binding logic.
