+++ app/Providers/AppServiceProvider.php
@@ -2,7 +2,14 @@
 
 namespace App\Providers;
 
+use App\Models\Media;
 use App\Models\User;
+use App\Models\Version;
+use Illuminate\Cache\RateLimiting\Limit;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\RateLimiter;
+use Illuminate\Support\Facades\Route;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Support\ServiceProvider;
 
@@ -25,6 +32,30 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot(): void
     {
+        $this->configureRateLimiting();
+
+        Route::bind('media', function (string $identifier): Media {
+            $user = Auth::user() ?? User::whereName(Route::getCurrentRoute()->parameter('user'))->firstOrFail();
+            return $user->Media()->whereIdentifier($identifier)->firstOrFail();
+        });
+
+        Route::bind('version', function (int $versionNumber): Version {
+            $media = Route::getCurrentRoute()->parameter('media');
+            return $media->Versions()->whereNumber($versionNumber)->firstOrFail();
+        });
+    }
+
+    /**
+     * Configure the rate limiters for the application.
+     *
+     * @return void
+     */
+    protected function configureRateLimiting(): void
+    {
+        RateLimiter::for('api', function (Request $request) {
+            return Limit::perMinute(60)->by($request->user()?->id ?: $request->ip());
+        });
+
         Relation::enforceMorphMap([
             'user' => User::class,
         ]);
