+++ config/app.php
@@ -174,8 +174,6 @@ return [
         App\Providers\AuthServiceProvider::class,
         // App\Providers\BroadcastServiceProvider::class,
         App\Providers\EventServiceProvider::class,
-        App\Providers\RouteServiceProvider::class,
-
         App\Providers\CdnHelperServiceProvider::class,
         App\Providers\CloudStorageServiceProvider::class,
         App\Providers\SqsFifoServiceProvider::class,
